chore: add missing or fix implicit transitive peerDependencies
Yarn 2 introduce an error for this during `yarn install`: YN0002 - MISSING_PEER_DEPENDENCY

For details see https://dev.to/arcanis/implicit-transitive-peer-dependencies-ed0

TL;DR: If you write a package that depends on Foo, and if Foo has a peer dependency, then you must provide it in either of the dependencies or peerDependencies fields. You won't "implicitly inherit" the peer dependencies declared in Foo.